// src/app/models/informacoes-rastreio.model.ts
export interface InformacoesRastreio {
	DTEMINOT: 	string; // - Data de Emissão da NFe
	DTEMICTR:	string; // - Data de Emissão do CTe
	DTENTCON: 	string; // - Data de SLA
	DTCOLETA: 	string; // - Data de Coleta
	DTPREENT:	string;	// - Data de Previsão de Entrega 
	DTENTREG:  	string; // - Data de Entrega
	DTBLOQUE:	string;
	DTDESBLO:	string;	
	TXCANHOT:	string;
	IDG043: string;
	DTSAICAR: string;
	TPDELIVE: number;
}